min change

+++ src/main/java/com/atomika/gitByCity/service/CommentService.java
@@ -3,8 +3,9 @@ package com.atomika.gitByCity.service;
 
 import com.atomika.gitByCity.dto.Comment;
 import com.atomika.gitByCity.dto.mapper.CommentMapper;
-import com.atomika.gitByCity.entity.CommentEntity;
+import com.atomika.gitByCity.repositories.ClientRepository;
 import com.atomika.gitByCity.repositories.CommentRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,25 +17,44 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final CommentMapper commentMapper;
+    private final ClientRepository clientRepository;
 
-    public Comment create (Comment comment) {
+
+    @Transactional
+    public Comment create (Comment comment, String clientName) {
+        comment.setClientId(clientRepository.findClientIdByUsername(clientName));
         return commentMapper.entityToDto(
                 commentRepository.save(commentMapper.dtoToEntity(comment)));
     }
 
-    public Comment update (Comment comment) {
+    public Comment update (Comment comment, String clientName) {
+        comment.setClientId(clientRepository.findClientIdByUsername(clientName));
         return commentMapper.entityToDto(
                 commentRepository.save(commentMapper.dtoToEntity(comment)));
     }
 
-    public Long delete (Long id) {
-        commentRepository.deleteById(id);
-        return id;
+
+    //todo посмотрет потом //ошбика была в том что на уровне связей стояло nullable = false и из за этого не давало удалить
+    @Transactional
+    public Long delete(Long id) {
+        if (commentRepository.existsById(id)) {
+            commentRepository.deleteById(id);
+            return id;
+        } else {
+            throw new RuntimeException("нет такого коммента" + id);
+        }
     }
 
-    public List<Comment> findAll () {
-        return commentMapper.toList(commentRepository.findAll());
+    @Transactional
+    public List<Comment> findAllForRoute (long routeId) {
+        return commentMapper.toList(commentRepository.findAllForRoute(routeId));
     }
+
+    @Transactional
+    public List<Comment> findAllForPoint (long pointId) {
+        return commentMapper.toList(commentRepository.findAllForPoint(pointId));
+    }
+
 
     public boolean isCreator (String username, long commentId) {
         return commentRepository.isCreator(username, commentId);
